dune: dune based build system
Based on Christian Lindig's work.

Initially this will be used to build unit tests, and to make development
easier.

Dune supports proper incremental builds and editor integration with
merlin/LSP.

For now the Makefile based build system is retained too: this is not a
hard dependency on Dune.

Using version 2.1 of Dune build language here, because that is the one
available in Ubuntu Focal (part of the CI here).

No functional change.

Co-authored-by: Edwin Török <edvin.torok@citrix.com>
Co-authored-by: Christian Lindig <christian.lindig@citrix.com>
Acked-by: Christian Lindig <christian.lindig@citrix.com>
Signed-off-by: Edwin Török <edvin.torok@citrix.com>
Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>