CP-52226 - oxenstored: Implement partial directory call for dom0 with per-connection cache
Implements XS_DIRECTORY_PART (originally specified in
https://lore.kernel.org/xen-devel/20161205074853.13268-1-jgross@suse.com/).
Like in CXenstored, it's limited to dom0. It is intended to list
directories larger than the size of a single payload (4096 bytes) over
several calls (not necessarily inside a single transaction).

To alleviate the impact of the differences in how node's children are
represented between C and OCaml Xenstoreds, implements a per-connection
cache allowing the string representation of node's children to be generated
once and reused later as long as the underlying list of children doesn't change
(and the connection is reused).

Whereas CXenstore uses a global generation count, modified at
transaction creation and node changes, the OCaml version relies on
comparisons of immutable trees. A per-node generation count is instead
faked in the connection cache - it allows clients to know that the
underlying information has changed and they should restart the sequence
of partial directory calls to get a consistent picture.
(i.e., while the generation count returned by CXenstore is globally unique,
this is not useful for this particular function, since generation count
is only compared against the results of previous invocations of partial
directory calls on the same node)

The cache is refreshed by OXenstored itself, and is limited to 8 entries
(by default, configurable).

A direct translation of the C code when it comes to string handling and
allocation.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>