tools/ocaml: add unit test skeleton with Dune build system
Based on initial work by Christian Lindig and Edwin Torok.

The unit tests do not require Xen to be available, so add mock
eventchn, gnt, plugin libraries for the unit test to use,
and copy the non-system specific modules from oxenstored/ to
tests/.

The gnt library requires some trickery, as it needs to return an opaque
Xenmmap.mmap_interface type - so create an unsafe_get casting 'unit' to
this type (it will not be used in testing, but we need to instantiate it.
The other option would be to mock the entire xenmmap library, which also
requires manipulating other libraries dependent on it - so this is the
easiest way).

Xenstored had to be split into Xenstored and Xenstored_main, so that we can
use the functions defined in Xenstored without actually starting up the
daemon in a unit test. Similarly argument parsing had to be delayed until
after daemon startup.

Add a Testing_status module with the 'under_testing' global that gets overriden
in tests/. It is passed to Xenbus.Xb.{open_mmap,close} by callers to
avoid doing things unavailable in the testing environment. It is also used to
disable setrlimit when running as non-root in poll.ml, and to disable search
for plugin implementation.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>
Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>