tests: Add unit tests for basic functionality
Add an 'unsafe_pop_output' function to Xenbus to inspect response packets
that are pushed into the queue for writing without actually writing them out
(which would be hard to get working in the testing environment).
The function asserts it runs as a non-root and should not be used
outside of unit tests.
(This is the least intrusive way of getting the unit testing loop
working)

These unit tests (and the ones in the following commits) are adaptations
of Mirage ocaml-xenstore unit tests.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>