tests: Add unit tests for watches
Several operations push the resulting Watchevent onto the queue first,
and only then push the response to the call itself. Due to the hacky way
tests are peeking into the queue, a few tests need to work around this
themselves.

A few tests in the Mirage ocaml-xenstore repo were left as TODOs with
names and descriptions only. I've implemented them here.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>